add `embedded-hal-async` driver
This branch adds an `embedded-hal-async` version of the Blackberry Q10
keyboard driver. The async driver is basically identical to the blocking
driver, except its methods are async, and it returns the error type
defined by the `embedded-hal-async::i2c::I2c` trait.

Because `embedded-hal-async` depends on nightly Rust features, the async
driver is behind an off-by-default feature flag.

I've also done some refactoring to the main (blocking) driver so that
parts of it can be more easily reused with the async driver.

Closes tosc-rs/mnemos#144